Make cinema and movie_info endpoints memoized functions
