+++ src/lib/utils.js
@@ -1,0 +1,13 @@
+export async function fetch_page(url, params = {}) {
+	try {
+		let response = await fetch(url, params);
+
+		if (!response.ok) {
+			throw new Error('Unable to fetch');
+		}
+
+		return response;
+	} catch {
+		throw new Error('Unable to fetch');
+	}
+}
